Extract common code fragments into new component
