+++ src/modules/main/sidebar/index.ts
@@ -1,3 +1,5 @@
 export * from './Sidebar';
 export * from './models';
-export * from './portal';
+export * from './sidebar-panel';
+export * from './sidebar-links';
+
