--- src/modules/main/sidebar/portal/Portal.scss
@@ -1,22 +1,0 @@
-@import 'styles/mixins';
-
-.portal {
-    @include col;
-    border-right: 1px solid rgba($primary, 0.43);
-    background: $surface;
-    position: absolute;
-    z-index: 6;
-    height: 100vh;
-    width: 200px;
-}
-
-.tile {
-    display: flex;
-
-    p {
-        vertical-align: middle;
-        display: inline-block;
-        margin-left: 5px;
-        text-transform: capitalize;
-    }
-}
