--- src/modules/main/sidebar/portal/index.ts
@@ -1,1 +1,0 @@
-export * from './Portal';
