+++ src/modules/main/sidebar/sidebar-links/index.ts
@@ -1,0 +1,1 @@
+export * from './SidebarLinks'
