+++ src/modules/main/sidebar/sidebar-panel/index.ts
@@ -1,0 +1,1 @@
+export * from './SidebarPanel';
